move restart button from Settings to Version page. only show Factory Reset when in developer mode
